Fixing Parser Unit Tests (#372)
Parser unit tests would always pass, regardless of whether the answer
was correct or not, due to incorrect pattern matching.
- Now each unit tests expects a concrete value produced by a given
parser (previously was always a string)
- Bumped the precision of our float literal parser from "Float" type to
"Double" type, as it was not precise enough for some of the test cases.
- There exists BigDecimal type in Java which is able to represent
arbitrary-precision decimal numbers, however I chose not to for now in
line with not overthinking it too much!

+++ core/src/main/scala-3/Parser.scala
@@ -8,7 +8,7 @@ import scair.clair.macros.DerivedOperationCompanion
 import scair.dialects.builtin.ModuleOp
 import scair.ir.*
 
-import java.lang.Float.parseFloat
+import java.lang.Double.parseDouble
 import java.lang.Math.pow
 import scala.annotation.switch
 import scala.annotation.tailrec
@@ -228,10 +228,17 @@ object Parser:
     P("0x" ~~ HexDigits.!).map((hex: String) => BigInt(hex, 16))
 
   private def parseFloatNum(float: (String, String)): Double =
-    val number = parseFloat(float._1).toDouble
-    val power = parseFloat(float._2).toDouble
+    val number = parseDouble(float._1)
+    val power = parseDouble(float._2)
     return number * pow(10, power)
 
+  /** Parses a floating-point number from its string representation. NOTE: This
+    * is only a float approximation, and in its current form should not be
+    * trusted for precision sensitive applications.
+    *
+    * @return
+    *   float: (String, String)
+    */
   def FloatLiteral[$: P] = P(
     (CharIn("\\-\\+").? ~~ DecDigits ~~ "." ~~ DecDigits).!
       ~~ (CharIn("eE")
